refactor(filter): extract reusable FilteredSubrequestExecutor from IRR
Extract the filter + transport sub-request lifecycle out of the
iterative_request_router into an owned, caller-agnostic
FilteredSubrequestExecutor under filter/src/filtered_subrequest/. The IRR
becomes a thin adapter that maps the executor's continuation types onto
its own step/transition vocabulary, so the rest of the router is
unchanged.

Strictly behavior-preserving (issue #1073): no config surface changes
(those land in a follow-up), the IRR remains the only consumer, and the
lifecycle unit tests move to the new module boundary. A bridge test pins
the TransportFailure -> config::TransportErrorKind mapping so
transport-error transition routing stays covered after the type split.

TerminalResponse handling is intentionally left as-is and deferred to a
focused follow-up.
