+++ crates/filter/src/builtins/http/traffic_management/iterative_request_router/mod.rs
@@ -16,7 +16,7 @@
 //!
 //! # Streaming support
 //!
-//! When a step's filters select [`SubRequestResponseMode::Streaming`],
+//! When a step's filters select [`SubRequestResponseMode::Streaming`](crate::SubRequestResponseMode::Streaming),
 //! IRR dispatches via `send_streaming()` instead of the buffered
 //! `execute()` path. Header-safe failover transitions run before that
 //! step exposes bytes. After clean EOF and response-body completion,
@@ -60,29 +60,22 @@ use std::{
 use async_trait::async_trait;
 use bytes::Bytes;
 use http::HeaderMap;
-use pingora_core::upstreams::peer::HttpPeer;
 use tracing::{debug, info, warn};
 
 use self::{
     config::IterativeRequestRouterConfig,
-    runner::{IrrStepRunner, OpenedStepKind, StepRuntime},
-    streaming::{IrrStreamingBody, IrrStreamingSession, ensure_combined_retained_limit},
+    runner::{IrrStepRunner, OpenedStepKind},
+    streaming::{IrrStreamingSession, ensure_combined_retained_limit, step_completion_from},
 };
 use crate::{
-    FilterEntry, FilterError, FilterPipeline, FilterRegistry, IterationState, StreamTermination, SubRequest,
-    SubRequestResponseMode, SubResponse,
+    FilterEntry, FilterError, FilterPipeline, FilterRegistry, IterationState, NextIterationBody, RequestExtensions,
+    StreamTermination, SubRequest, SubResponse,
     actions::{FilterAction, Rejection, StreamingResponseBody as _, StreamingTerminalResponse, TerminalResponse},
     factory::parse_filter_config,
     filter::{HttpFilter, HttpFilterContext},
+    filtered_subrequest::{FilteredStreamingBody, SubrequestRuntime, normalize_response_status},
     pipeline::subrequest::DEPTH_HEADER,
 };
-
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-
-/// Streaming idle timeout (max time between chunks).
-const STREAMING_IDLE_TIMEOUT: Duration = Duration::from_secs(30);
 
 // ---------------------------------------------------------------------------
 // Utilities
@@ -117,6 +110,20 @@ pub(super) fn streaming_transition_order_is_valid(transitions: &[config::StepTra
             true
         }
     })
+}
+
+/// Strip the router-owned iteration extensions before handing state back to the
+/// parent request context.
+///
+/// The generic executor already removes its own transient mechanisms; the
+/// router additionally injects [`IterationState`] and [`NextIterationBody`],
+/// which must not leak into the parent pipeline. Applying this to the executor's
+/// parent-facing extensions restores the exact end state of the pre-extraction
+/// router.
+pub(super) fn strip_iteration_extensions(mut extensions: RequestExtensions) -> RequestExtensions {
+    extensions.remove::<IterationState>();
+    extensions.remove::<NextIterationBody>();
+    extensions
 }
 
 // ---------------------------------------------------------------------------
@@ -428,7 +435,7 @@ impl IterativeRequestRouterFilter {
             depth,
             max_response_bytes,
             self.max_state_bytes,
-            StepRuntime {
+            SubrequestRuntime {
                 client_addr: ctx.client_addr,
                 downstream_tls: ctx.downstream_tls,
                 peer_identity: ctx.peer_identity.clone(),
@@ -482,7 +489,7 @@ impl IterativeRequestRouterFilter {
                     let transitions = self.step_transitions.get(&current_step).map_or(&[][..], Vec::as_slice);
                     if !streaming_transition_order_is_valid(transitions) {
                         (*body).cancel().await;
-                        ctx.extensions = continuation.into_parent_extensions();
+                        ctx.extensions = strip_iteration_extensions(continuation.into_parent_extensions());
                         return Err(format!(
                             "iterative_request_router: step '{current_step}' selected streaming with interleaved transition phases"
                         )
@@ -495,19 +502,21 @@ impl IterativeRequestRouterFilter {
                                 to = next_step.as_ref(),
                                 "streaming header failover before response commitment"
                             );
-                            let mut skipped = IrrStreamingBody::new(body, continuation);
+                            let mut skipped = FilteredStreamingBody::new(body, continuation);
                             if let Err(error) = skipped.suppress().await {
-                                ctx.extensions = skipped.into_continuation().into_parent_extensions();
+                                ctx.extensions =
+                                    strip_iteration_extensions(skipped.into_continuation().into_parent_extensions());
                                 return Err(error);
                             }
-                            let mut completion = match skipped.into_continuation().into_completion() {
-                                Ok(completion) => completion,
-                                Err(error) => {
-                                    let (error, restored_extensions) = error.into_parts();
-                                    ctx.extensions = restored_extensions;
-                                    return Err(error);
-                                },
-                            };
+                            let mut completion =
+                                match step_completion_from(skipped.into_continuation().into_completion()) {
+                                    Ok(completion) => completion,
+                                    Err(error) => {
+                                        let (error, restored_extensions) = error.into_parts();
+                                        ctx.extensions = restored_extensions;
+                                        return Err(error);
+                                    },
+                                };
                             completion.state.previous_response = None;
                             completion.state.iteration += 1;
                             if completion.state.retained_bytes() > self.max_state_bytes {
@@ -528,9 +537,9 @@ impl IterativeRequestRouterFilter {
                             current_step = next_step;
                         },
                         TransitionResult::Done | TransitionResult::NoMatch => {
-                            let Some(active_state) = continuation.extensions.get::<IterationState>() else {
+                            let Some(active_state) = continuation.extensions().get::<IterationState>() else {
                                 (*body).cancel().await;
-                                ctx.extensions = continuation.into_parent_extensions();
+                                ctx.extensions = strip_iteration_extensions(continuation.into_parent_extensions());
                                 return Err(
                                     "iterative_request_router: iteration state missing before stream handoff"
                                         .to_owned()
@@ -545,7 +554,7 @@ impl IterativeRequestRouterFilter {
                             .is_err()
                             {
                                 (*body).cancel().await;
-                                let completion = match continuation.into_completion() {
+                                let completion = match step_completion_from(continuation.into_completion()) {
                                     Ok(completion) => completion,
                                     Err(error) => {
                                         let (error, restored_extensions) = error.into_parts();
@@ -579,7 +588,7 @@ impl IterativeRequestRouterFilter {
                     }
                 },
                 OpenedStepKind::Complete(mut outcome) => {
-                    let completion = match continuation.into_completion() {
+                    let completion = match step_completion_from(continuation.into_completion()) {
                         Ok(completion) => completion,
                         Err(error) => {
                             let (error, restored_extensions) = error.into_parts();
@@ -744,54 +753,6 @@ pub(super) enum TransitionResult {
     NoMatch,
 }
 
-/// Convert transport failures into a gateway status and error classification.
-pub(super) fn classify_transport_failure(
-    error: &praxis_core::subrequest::SubRequestError,
-) -> (u16, config::TransportErrorKind) {
-    use praxis_core::subrequest::SubRequestError;
-    match error {
-        SubRequestError::AdmissionTimeout { .. } => (503, config::TransportErrorKind::AdmissionTimeout),
-        SubRequestError::CircuitOpen { .. } => (503, config::TransportErrorKind::CircuitOpen),
-        SubRequestError::Connect(_) => (502, config::TransportErrorKind::Connect),
-        SubRequestError::DeadlineExceeded => (504, config::TransportErrorKind::DeadlineExceeded),
-        SubRequestError::ResponseTooLarge { .. } => (502, config::TransportErrorKind::ResponseTooLarge),
-        _ => (502, config::TransportErrorKind::Io),
-    }
-}
-
-/// Convert a nested filter's local response into transition input.
-fn subresponse_from_rejection(rejection: Rejection) -> SubResponse {
-    let status = normalize_response_status(rejection.status);
-    let mut headers = HeaderMap::new();
-    for (name, value) in rejection.headers {
-        let Ok(name) = http::HeaderName::try_from(name) else {
-            continue;
-        };
-        let Ok(value) = http::HeaderValue::try_from(value) else {
-            continue;
-        };
-        headers.append(name, value);
-    }
-    if let Some(header_map) = rejection.header_map {
-        for (name, value) in header_map.iter() {
-            headers.append(name.clone(), value.clone());
-        }
-    }
-    let mut response = SubResponse {
-        status,
-        headers,
-        body: rejection.body.unwrap_or_default(),
-    };
-    sanitize_subresponse_headers(&mut response.headers);
-    response
-}
-
-/// Keep final locally generated statuses inside the terminal response range.
-/// Informational, invalid upstream, or invalid custom-filter values become 502.
-fn normalize_response_status(status: u16) -> u16 {
-    if (200..=599).contains(&status) { status } else { 502 }
-}
-
 /// Evaluate transition rules against a step outcome.
 pub(super) fn evaluate_transitions(
     transitions: &[config::StepTransition],
@@ -877,156 +838,6 @@ fn parse_depth(request: &crate::Request) -> u8 {
         .unwrap_or(0)
 }
 
-/// Strip all reserved internal headers from sub-request headers
-/// so the core executor re-injects depth via
-/// [`FrameworkHeaders`](praxis_core::subrequest::FrameworkHeaders).
-///
-/// The depth header uses a reserved `x-praxis-*` prefix, so it
-/// is covered by the [`is_reserved`] check.
-///
-/// [`is_reserved`]: praxis_core::reserved_headers::is_reserved
-fn strip_reserved_headers(headers: &mut HeaderMap) {
-    let to_remove: Vec<http::header::HeaderName> = headers
-        .keys()
-        .filter(|name| praxis_core::reserved_headers::is_reserved(name.as_str()))
-        .cloned()
-        .collect();
-    for name in to_remove {
-        headers.remove(&name);
-    }
-}
-
-/// Apply request mutations emitted across the step's header and body
-/// filter phases before dispatching its upstream request.
-fn apply_request_header_mutations(headers: &mut HeaderMap, ctx: &HttpFilterContext<'_>) {
-    for name in &ctx.request_headers_to_remove {
-        headers.remove(name);
-    }
-    for (name, value) in &ctx.request_headers_to_set {
-        headers.insert(name.clone(), value.clone());
-    }
-    for (name, value) in &ctx.extra_request_headers {
-        if let (Ok(name), Ok(value)) = (
-            http::header::HeaderName::from_bytes(name.as_bytes()),
-            http::HeaderValue::from_str(value),
-        ) {
-            headers.insert(name, value);
-        }
-    }
-}
-
-/// Apply body pre-read mutations to the request snapshot that header
-/// filters use for classification and routing.
-fn apply_pre_read_header_mutations(headers: &mut HeaderMap, ctx: &HttpFilterContext<'_>) {
-    if ctx.pre_read_mutations.is_empty() {
-        apply_request_header_mutations(headers, ctx);
-        return;
-    }
-
-    for mutation in &ctx.pre_read_mutations {
-        match mutation {
-            crate::TrustedHeaderMutation::Remove(name) => {
-                headers.remove(name);
-            },
-            crate::TrustedHeaderMutation::Set(name, value) => {
-                headers.insert(name.clone(), value.clone());
-            },
-            crate::TrustedHeaderMutation::Add(name, value) => {
-                if let Ok(value) = http::HeaderValue::from_str(value) {
-                    headers.append(name.clone(), value);
-                }
-            },
-        }
-    }
-}
-
-/// Remove inbound message-framing headers after request-body filters
-/// have potentially changed the payload. The subrequest executor adds
-/// the correct `Content-Length` for non-empty bodies.
-fn strip_request_framing_headers(headers: &mut HeaderMap) {
-    headers.remove(http::header::CONTENT_LENGTH);
-    headers.remove(http::header::TRANSFER_ENCODING);
-}
-
-/// Headers that apply only to one HTTP connection and must not cross it.
-const REQUEST_HOP_BY_HOP: &[&str] = &[
-    "connection",
-    "keep-alive",
-    "proxy-authenticate",
-    "proxy-authorization",
-    "te",
-    "trailer",
-    "transfer-encoding",
-    "upgrade",
-];
-
-/// Response-side hop-by-hop headers.
-const RESPONSE_HOP_BY_HOP: &[&str] = &[
-    "connection",
-    "keep-alive",
-    "proxy-authenticate",
-    "te",
-    "trailer",
-    "transfer-encoding",
-    "upgrade",
-];
-
-/// Apply the same forwarding boundary as the normal upstream path.
-fn sanitize_subrequest_headers(headers: &mut HeaderMap) {
-    strip_hop_by_hop_headers(headers, REQUEST_HOP_BY_HOP);
-    strip_reserved_headers(headers);
-    strip_request_framing_headers(headers);
-}
-
-/// Supply the selected upstream authority without carrying a prior step's Host.
-fn ensure_destination_host(headers: &mut HeaderMap, address: &str) -> Result<(), FilterError> {
-    if !headers.contains_key(http::header::HOST) {
-        let value = http::HeaderValue::from_str(address).map_err(|error| -> FilterError {
-            format!("iterative_request_router: invalid upstream Host: {error}").into()
-        })?;
-        headers.insert(http::header::HOST, value);
-    }
-    Ok(())
-}
-
-/// Remove connection-scoped and proxy-internal response metadata.
-fn sanitize_subresponse_headers(headers: &mut HeaderMap) {
-    strip_hop_by_hop_headers(headers, RESPONSE_HOP_BY_HOP);
-    strip_reserved_headers(headers);
-}
-
-/// Remove the static hop-by-hop set and headers named by `Connection`.
-fn strip_hop_by_hop_headers(headers: &mut HeaderMap, static_headers: &[&str]) {
-    let connection_values: Vec<_> = headers.get_all(http::header::CONNECTION).iter().cloned().collect();
-    for name in static_headers {
-        headers.remove(*name);
-    }
-    for value in connection_values {
-        let Ok(value) = value.to_str() else { continue };
-        for token in value.split(',').map(str::trim).filter(|token| !token.is_empty()) {
-            headers.remove(token);
-        }
-    }
-}
-
-/// Whether a fully buffered nested body exceeds its pipeline mode's
-/// configured ceiling.
-fn body_exceeds_limit(mode: crate::body::BodyMode, body_len: usize) -> bool {
-    match mode {
-        crate::body::BodyMode::SizeLimit { max_bytes }
-        | crate::body::BodyMode::StreamBuffer {
-            max_bytes: Some(max_bytes),
-        } => body_len > max_bytes,
-        crate::body::BodyMode::Stream | crate::body::BodyMode::StreamBuffer { max_bytes: None } => false,
-    }
-}
-
-/// Whether a response exceeds either the iterative router's global
-/// per-step ceiling or the nested pipeline's body-mode ceiling.
-fn response_body_exceeds_limits(mode: crate::body::BodyMode, max_response_bytes: usize, body_len: usize) -> bool {
-    body_len > max_response_bytes || body_exceeds_limit(mode, body_len)
-}
-
 /// Clamp the router-specific response cap to the listener's global body mode.
 fn effective_response_limit(configured: usize, parent_mode: crate::body::BodyMode) -> usize {
     match parent_mode {
@@ -1036,151 +847,6 @@ fn effective_response_limit(configured: usize, parent_mode: crate::body::BodyMod
         } => configured.min(max_bytes),
         crate::body::BodyMode::Stream | crate::body::BodyMode::StreamBuffer { max_bytes: None } => configured,
     }
-}
-
-/// Extract only the listener-level streaming ceiling from a nested pipeline.
-///
-/// The IRR `max_response_bytes` setting is intentionally buffered-only. A
-/// nested `SizeLimit` is produced by listener body-limit propagation and must
-/// still constrain the live transport.
-fn streaming_transport_limit(mode: crate::body::BodyMode) -> Option<usize> {
-    match mode {
-        crate::body::BodyMode::SizeLimit { max_bytes } => Some(max_bytes),
-        crate::body::BodyMode::Stream | crate::body::BodyMode::StreamBuffer { .. } => None,
-    }
-}
-
-/// Whether a step filter grew the shared iteration state past its ceiling.
-fn iteration_state_exceeds_limit(ctx: &HttpFilterContext<'_>, max_state_bytes: usize) -> bool {
-    ctx.extensions
-        .get::<IterationState>()
-        .is_some_and(|state| state.retained_bytes() > max_state_bytes)
-}
-
-/// Runtime resources inherited from the pipeline that contains the
-/// iterative router.
-#[derive(Clone, Copy)]
-struct SubPipelineRuntimeResources<'a> {
-    /// Original downstream client address.
-    client_addr: Option<std::net::IpAddr>,
-
-    /// Whether the original downstream connection uses TLS.
-    downstream_tls: bool,
-
-    /// Shared endpoint-health state.
-    health_registry: Option<&'a praxis_core::health::HealthRegistry>,
-
-    /// Shared request ID generator.
-    id_generator: &'a praxis_core::id::IdGenerator,
-
-    /// Named runtime key-value stores.
-    kv_stores: Option<&'a praxis_core::kv::KvStoreRegistry>,
-
-    /// Verified downstream mTLS identity.
-    peer_identity: Option<&'a Arc<praxis_tls::TlsPeerIdentity>>,
-
-    /// Start time of the containing client request.
-    request_start: Instant,
-
-    /// Shared client used for recursive subrequests.
-    subrequest_client: Option<&'a praxis_core::subrequest::SubRequestClient>,
-
-    /// Server-provided wall-clock source.
-    time_source: &'a dyn praxis_core::time::TimeSource,
-}
-
-/// Build a [`HttpFilterContext`] for running a step's sub-pipeline,
-/// inheriting server-injected resources from the containing request.
-#[expect(clippy::too_many_lines, reason = "all fields must be initialized")]
-fn build_sub_filter_context<'a>(
-    pipeline: &'a FilterPipeline,
-    request: &'a crate::Request,
-    runtime: SubPipelineRuntimeResources<'a>,
-) -> HttpFilterContext<'a> {
-    HttpFilterContext {
-        buffered_request_body: None,
-        body_done_indices: Vec::new(),
-        branch_iterations: HashMap::new(),
-        client_addr: runtime.client_addr,
-        cluster: None,
-        current_filter_id: None,
-        downstream_tls: runtime.downstream_tls,
-        extensions: crate::extensions::RequestExtensions::default(),
-        executed_filter_indices: Vec::new(),
-        extra_request_headers: Vec::new(),
-        filter_metadata: HashMap::new(),
-        filter_results: HashMap::new(),
-        filter_state: HashMap::new(),
-        health_registry: runtime.health_registry,
-        id_generator: runtime.id_generator,
-        kv_stores: runtime.kv_stores,
-        session_stores: None,
-        metrics_route: None,
-        peer_identity: runtime.peer_identity.cloned(),
-        prior_pre_read_mutations: Vec::new(),
-        pre_read_mutations: Vec::new(),
-        request,
-        request_body_bytes: 0,
-        request_body_mode: pipeline.body_capabilities().request_body_mode,
-        request_headers_to_remove: Vec::new(),
-        request_headers_to_set: Vec::new(),
-        request_start: runtime.request_start,
-        response_body_bytes: 0,
-        response_body_mode: pipeline.body_capabilities().response_body_mode,
-        response_header: None,
-        response_headers_modified: false,
-        rewritten_path: None,
-        selected_endpoint_index: None,
-        attempted_endpoints: Vec::new(),
-        retry_policy: None,
-        route_retry_policy: None,
-        cluster_retry_state: None,
-        cluster_retry_state_released: false,
-        endpoint_reselector: None,
-        pinned_endpoint_address: None,
-        structured_metadata: HashMap::new(),
-        subrequest_client: runtime.subrequest_client,
-        subrequest_response_mode: SubRequestResponseMode::Buffered,
-        time_source: runtime.time_source,
-        upstream: None,
-    }
-}
-
-/// Convert a Praxis [`Upstream`] to a Pingora [`HttpPeer`].
-///
-/// Applies TLS settings (CA, client cert, verify toggle) and
-/// connection options (timeouts) from the upstream config. Derives
-/// SNI from the address hostname when not explicitly configured.
-///
-/// [`Upstream`]: praxis_core::connectivity::Upstream
-async fn build_peer(
-    upstream: &praxis_core::connectivity::Upstream,
-) -> Result<HttpPeer, praxis_core::connectivity::peer::AddressResolutionError> {
-    use praxis_core::connectivity::peer as peer_utils;
-
-    let addr: &str = &upstream.address;
-    let socket_addr = peer_utils::resolve_address(addr).await?;
-    let tls_enabled = upstream.tls.is_some();
-    let sni = upstream
-        .tls
-        .as_ref()
-        .and_then(|t| t.sni().map(str::to_owned))
-        .unwrap_or_else(|| {
-            if tls_enabled {
-                peer_utils::derive_sni(addr)
-            } else {
-                String::new()
-            }
-        });
-
-    let mut peer = HttpPeer::new(socket_addr, tls_enabled, sni);
-    peer_utils::apply_connection_options(&mut peer, &upstream.connection);
-
-    if let Some(tls) = &upstream.tls {
-        peer_utils::apply_cached_tls(&mut peer, tls, addr);
-    }
-
-    Ok(peer)
 }
 
 /// Build a [`TerminalResponse`] carrying the sub-request response.
